added theme

+++ src/app/api/buys/delete-batch/route.ts
@@ -4,9 +4,12 @@ import { NextRequest, NextResponse } from "next/server";
 import { deleteBuys } from "../services";
 
 export async function POST(request: NextRequest) {
-  const token = tokenValidator(request);
-  if (!token) {
-    return NextResponse.json("Solicitud no permitida", { status: 400 });
+  const token = await tokenValidator(request);
+  if (!token || token.role == "CASHIER") {
+    return NextResponse.json(
+      { message: "Solicitud no permitida" },
+      { status: 400 }
+    );
   }
   const body = await request.json();
   return await deleteBuys(body);
